requirements.txt
asgiref==3.6.0
Django~=4.1.6
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3

pip~=22.3.1
wheel~=0.37.1
Pillow~=9.4.0
Jinja2~=3.1.2
setuptools~=65.5.1
django-email-users~=1.1.0